Updated the seeds and the main map
The main map was not displaying all of the campgrounds in the last update. It was only showing the campgrounds that were being paginated. Fixed that in this update. Also updated the seeds so that it can be used on the live database, which was not happening previously as its connections string was setup differently.